refactor(nonce): implement nonce increment without libsodium

+++ tox_core/src/net_crypto/mod.rs
@@ -824,7 +824,7 @@ impl NetCrypto {
         let cur_last_bytes = CryptoData::nonce_last_bytes(received_nonce);
         let (diff, _) = packet.nonce_last_bytes.overflowing_sub(cur_last_bytes);
         let mut packet_nonce = received_nonce;
-        increment_nonce_number(&mut packet_nonce, u64::from(diff));
+        increment_nonce_number(&mut packet_nonce, diff);
 
         let payload = match packet.get_payload(&session_precomputed_key, &packet_nonce) {
             Ok(payload) => payload,
@@ -858,7 +858,7 @@ impl NetCrypto {
 
         // Update nonce if diff is big enough
         if diff > NONCE_DIFF_THRESHOLD * 2 {
-            increment_nonce_number(&mut received_nonce, u64::from(NONCE_DIFF_THRESHOLD));
+            increment_nonce_number(&mut received_nonce, NONCE_DIFF_THRESHOLD);
         }
 
         self.send_connection_status(&connection, true).await
@@ -2593,7 +2593,7 @@ mod tests {
 
         // Make the diff between nonces is bigger than the threshold
         let mut packet_nonce = received_nonce;
-        increment_nonce_number(&mut packet_nonce, u64::from(2 * NONCE_DIFF_THRESHOLD + 1));
+        increment_nonce_number(&mut packet_nonce, 2 * NONCE_DIFF_THRESHOLD + 1);
 
         let crypto_data_payload = CryptoDataPayload {
             buffer_start: 0,
@@ -2607,7 +2607,7 @@ mod tests {
         // The diff between nonces is bigger than the threshold so received
         // nonce should be changed increased
         let mut expected_nonce = received_nonce;
-        increment_nonce_number(&mut expected_nonce, u64::from(NONCE_DIFF_THRESHOLD));
+        increment_nonce_number(&mut expected_nonce, NONCE_DIFF_THRESHOLD);
         assert_eq!(unpack!(connection.status, ConnectionStatus::Established, received_nonce), expected_nonce);
 
         assert_eq!(connection.recv_array.buffer_start, 0);
